Create PositionFaces.py
Need to create 10k individual layers with face positioned on the Canvas exactly as it would be in the final artwork. All individual elements are their own individual layer files. They will be mixed, matched and merged into one artwork using a JS script. Hence, each element has to be positioned in its proper place on the final canvas. Which is why the 10k cropped heads which were retrieved using the slicing script, will need to be positioned on the canvas in their correct places (just above the punk body). This script could have been combined with the slicing script itself. I just like to chop my problems down in smaller steps :)